Configure the workspace on initialize request not only on DidChangeConfiguration
This prevents a crash when trying to invoke sourcery with implicitely unwrapped
optional attributes. I had the issue with a LSP client which was not calling
this event (FYI lsp-mode for emacs).
There is not absolute need to wait for the DidChangeConfiguration event to parse
the Package.swift but it will still be done if the workspace configuration
changes.